requirements.txt
pandas==2.1.4
pyarrow==16.1.0
fastparquet==2024.11.0
SQLAlchemy==1.4.52
psycopg2-binary==2.9.9
google-cloud-storage==2.17.0
google-cloud-bigquery==3.20.1
apache-airflow-providers-google==10.21.0